backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "stratos-database",
    "boto3>=1.40.29",
    "langfuse>=3.3.4",
    "openai-agents>=0.3.0",
    "pydantic-ai>=1.0.6",
    "python-dotenv>=1.1.1",
]

[tool.uv.workspace]
members = [
    "database",
    "api",
    "scheduler",
]

[tool.uv.sources]
stratos-database = { workspace = true }
